Update to Debug Protocol 1.36.x:
- Introduces a capability completionTriggerCharacters for announcing
the characters that a frontend UI should use to trigger completion.
- Adds an optional sortText attribute to the CompletionItem type.
- In the schema change “number” to “integer” in cases where a float
makes no sense, e.g. for IDs.
- Fixed type error in EvaluateResponse.memoryReference - now a String
instead of Long to match protocol.

+++ org.eclipse.lsp4j.debug/src/main/xtend-gen/org/eclipse/lsp4j/debug/OutputEventArguments.java
@@ -39,11 +39,12 @@ public class OutputEventArguments {
   
   /**
    * If an attribute 'variablesReference' exists and its value is > 0, the output contains objects which can be
-   * retrieved by passing 'variablesReference' to the 'variables' request.
+   * retrieved by passing 'variablesReference' to the 'variables' request. The value should be less than or
+   * equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  private Long variablesReference;
+  private Integer variablesReference;
   
   /**
    * An optional source location where the output was produced.
@@ -115,22 +116,24 @@ public class OutputEventArguments {
   
   /**
    * If an attribute 'variablesReference' exists and its value is > 0, the output contains objects which can be
-   * retrieved by passing 'variablesReference' to the 'variables' request.
+   * retrieved by passing 'variablesReference' to the 'variables' request. The value should be less than or
+   * equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
   @Pure
-  public Long getVariablesReference() {
+  public Integer getVariablesReference() {
     return this.variablesReference;
   }
   
   /**
    * If an attribute 'variablesReference' exists and its value is > 0, the output contains objects which can be
-   * retrieved by passing 'variablesReference' to the 'variables' request.
+   * retrieved by passing 'variablesReference' to the 'variables' request. The value should be less than or
+   * equal to 2147483647 (2^31 - 1).
    * <p>
    * This is an optional property.
    */
-  public void setVariablesReference(final Long variablesReference) {
+  public void setVariablesReference(final Integer variablesReference) {
     this.variablesReference = variablesReference;
   }
